requirements.txt
# Python Dependencies for Twitter Sentiment Analysis
# =================================================

# Kafka Producer (kafka-python-ng for Python 3.12+ compatibility)
kafka-python-ng>=2.2.0

# Legacy kafka-python (for Python < 3.12)
# kafka-python==2.0.2

# Required for PySpark on Python 3.12+
setuptools>=69.0.0

# Apache Spark (PySpark)
pyspark==3.5.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Avro support
fastavro==1.9.0
avro-python3==1.10.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

